+++ CHANGELOG.md
@@ -14,7 +14,9 @@ All notable changes to this project will be documented in this file.
 - Continuous Integration & Continuous Delivery through Github Actions to Pypi
 - Add Microsoft Visual Studio Compiler (MSVC) compatibility
 - Uses Clang++ as a default compiler on Windows when execute `make`
-- `Helper Class`: Add .frequencies2cents() method
+- `Helper Class`:
+  - Add .frequencies2cents() method
+  - Add .freq2equalTemperament() method
 - `Interval Class`: Add .toCents() method
 - `Chord Class`: Add .toCents() methods
 - Add `Score::getComposerName()` method